Pull congress from GitHub to address CI failures after deprecation
Congress is no longer included in the CI image builds, which means
that when the opendev URL is translated to a local file path in
'get-ansible-role-requirements.yml' CI builds for stable branches
such as Ussuri fail.

This patch aims to pull the congress source from GitHub's archive
copy in order to force the CI jobs to pull from the remote.

Change-Id: I85089d9d1c63d1585e02da2ccaadbed20f356ae0