Add user defined collections
This patch allows the deployer to create a file which adds to, or
overrides the ansible collections specified in the openstack-ansible
repo ansible-collection-requirements.yml file.

The file /etc/openstack_deploy/user-collection-requirements.yml is
read during installation of ansible collections with the bootstrap_ansible.sh
script. Any extra collections listed will be installed and will take
precedence over collections of the same name specified in
ansible-collection-requirements.yml

Depends-On: https://review.opendev.org/754706
Change-Id: I22b55291c6898312a26cc5dbf7a3ae2f327429bd