Use nodepool epel mirror in CI for systemd-networkd package
This is failing often in CI and we should use the mirror when
possible.

The normal variable centos_epel_mirror is not defined in an
openstack-ansible user variable until after the AIO host prep
has happened, so we must detect the nodepool mirror directly
when doing the initial host preparation.

This patch refactors gathering the nodepool vars completely, to
read them into a new ansible var nodepool_vars where they can be
accessed in the whole bootstrap_host role.

Depends-On: https://review.opendev.org/755065
Change-Id: Ida3762ffff2140b3e5b287443004fa1f16addd53