Update variables for default Zun deployments and AIO
The Zun API added the ability to delete networks in this
microversion which means that tempest tests can be run without the
need to bind the Docker daemon to a wider set of addresses than
localhost.

For the same reason, one tempest test is skipped as this assumes
the Docker daemon is accessible on localhost.

This commit also enables the Zun Horizon UI when Zun is used

Change-Id: Ie41c08f3e8c8c06a8bd334f7d8ec9bf03586ec2b