Add to bind-to-mgmt upgrade release note for older deployments
Having upgraded an older deployment through to Victoria we
encountered an issue with the Ansible RabbitMQ modules and the
rabbitmqctl tool following the bind-to-mgmt config changes. This
was caused by the presence of old '127.0.1.1' host records which
are no longer added or managed by OSA.

This release note adds an extra word of warning for those with
long-standing OSA deployments.

Change-Id: Ife900f813b0282bf61ae6ce74a7d4c916ea8664e