Use zuul provided requirements repo during bootstrap-ansible.sh
This patch automatically detects if the deployment is running in
zuul CI  and uses the locally available git repo to get a copy
of upper-constraints.txt if possible.

A copy of the upper-constraints.txt file is placed in /opt for
all deployments both in and out of CI, and is used as the source
of constraints when creating the ansible runtime venv.

This reduces the need to make requests for the same data to the
opendev.org git servers which will improve CI job reliability.

Change-Id: Ib5afc1d6b46790f1ae41593f0ce8f8530320472e