Trigger deploy guide rebuild on a-r-r update
In deploy guide we rely on latest_tag to provide users with valid git tag
in the guide. However this is broken now, since deploy guide rebuild
is not triggered by jobs.

This change makes use of long ago added project-template that
also triggers build on a-r-r update comparing to it's original [1]

[1] https://opendev.org/openstack/openstack-zuul-jobs/src/commit/9f64b09a4b9dee2c5adcde696b59c689237c5d7a/zuul.d/project-templates.yaml#L1628-L1649

Change-Id: I4708b2ad183b4c09582f723d6c114012c8a6ea17