Gather minimal facts for dynamic_address_fact
Currently this code assumes that there are facts available for all
network interfaces on the host. This may be hundreds of interfaces
for compute or network nodes.

If the OSA fact gathering is switched to be minimal to increase
deployment performance then is is necessary to gather information
about the interface to take the dynamiclly selected IP address from.

This patch adds a network facts gathering task which is filtered to
return only facts about the network interfaces specified in
container_networks, or in a new override dynamic_address_gather_filter.

The network facts are also now accessed via ansible_facts[] instead
of the hostvar ansible_<interface>.

Change-Id: Ib6c117b27b84d911baad8f85ec3128965dd4c6c2