Gather minimal facts in common playbooks
The common playbooks for cinder, neutron and nova have an initial
facts gathering step.

This patch switches from implicit gather_facts to direct use of
the ansible setup module. The minimum set of facts are gathered,
and an additional filtered set of facts about the target processor
count is gathered. The processor count is used in many service
templates to set the number or threads/processes used.

This approach significantly reduces the size of the gathered
facts which cannot be achieved directly with a gather_subset
of '!all,min,hardware'.

In addition, all uses of ansible facts are converted to
ansible_facts[].

Change-Id: Id29f7344da7f71c9513d62c939ce2baead6ef451