Remove keepalived haproxy notification script
This script triggers a restart of HAProxy when any keepalived
instance enters the fault state. This can cause an unnecessary
restart of HAProxy, dropping internal connections, if the external
instance fails.

An alternative approach would be to associate this notification
with the internal instance only, but given the nature of the
existing HAProxy liveness check and the use of systemd across all
deployments this appears to be covered by service restart policy.

Change-Id: Ia1b7a3d53e6a3be064acabb6c6dd2e921c85ca94