Gather additional required facts to min
With minimal facts gathering we will experience failures while trying to
evaluate amount of threads for apis or mounts for hardening. So to avoid
gathering full hardware subset, we add common-task which will gather
only specific subset that can be additionally filtered.

Gathering processor or mounts subsent simply does not work with ansible
and result in full hardware subset as well.

Change-Id: Ia5802b4ec0b18271b8c5fbcc5574b484c5233a01