Increase backend_fall value for haproxy_letsencrypt_service
Currently to get LetsEncrypt certificate, we spawn python http server for 5 seconds to get haproxy endpoint up and after it's terminated, we run certbot which is a backend for this endpoint.

With current haproxy backend settings(inter=2000,fall=2), certbot has only 4 seconds for certificate validation which is too short in many cases.

Changing backend_fall from 2 to 5 should solve the problem.

Change-Id: I0bf2fa4f9f0206264984d50376de8aa8396e3d09