Add shallow_since to parallel git clone
shallow_since controls from which date history repo will be shallowed.
It has prescedence over depth, so when shallow_since is set for the repo
depth won't be used. This is handy
for some repos, like ceph-ansible, that
merge a lot of changes in bunches, which results in bootstrap issues.

Later, shallow_since will be managed with osa CLI and bump will
update this date for the repos in a-r-r.

Change-Id: I8017a01853e8f8950e6e8c7c2337e384af8d7689
(cherry picked from commit ed035fb275f8a296bfe29918931cbde2ff6da438)