haproxy: decrease check interval for letsencrypt back ends
When HTTP challenges are used, the HAProxy letsencrypt integration
allows a short timeout for the back end to be recognised as up
before the challenge is performed. The role default for this is 5
seconds.

The default interval to check an HAProxy back end is 12 seconds,
which means that the challenge response can frequently return
an HTTP 503 error. This change reduces the check interval to be
below 5 seconds to improve this reliability.

Change-Id: I1297965a2031eb49b4f048e09a2a6bb119921e24