Fix permissions for files created on repo server
If wheels or requirements files are created on a secondary
repo server these cannot be modified/deleted by the primary
repo server when it comes back online as this uses rsync as the
'nginx' user. This patch allows the owner of the created files
to be set explicitly to avoid this issue.

Depends-On: https://review.opendev.org/c/openstack/ansible-role-python_venv_build/+/804073
Change-Id: Ifdbac876013068855cd4405ca1bbacd262ec4d74