Add config_tempalte to collections
Since config_tempalte is really a handy module that can be used
outside of the OSA, we make it installable as collection and
start using it as a collection as well. This will also bring
us unambiguity about what exact module we use, since ceph-ansible
have their own interpritation of config_template that they ship.

Change-Id: Id52d56e387056066bfce28c781b24e8d98361ac5