Reduce ceph memory overhead for AIO by setting is_hci to true
This reserves 80% of the host memory to *not* be used when calculating
memory limits for the ceph osd processes. This setting is used in
ceph-ansible for 'hyperconverged' deployments where nova compute is
co-located with osd on the same nodes and the majority of memory must
be available for vm guests rather than storage.

Change-Id: I7bba3429d71b30253bc72fd013cec4d085eb8fbf