Use local repositories for ansible collections in CI
This patch will search for repositories in the zuul git cache which
match the names of ansible collections, for example
'openstack/ansible-collections-openstack`, and rewrite the collections
requirement file to point to the local copies.

This reduces the number of external git clones, and permits the use
of depends-on to validate changes to collections.

Depends-On: Ice2dbac8a2b23e33523e9b6f120f8f16588b43e1
Change-Id: I2ba12ece8e765a6c409d50c95f4cfbe1f314180f