Add defaults for TLS 1.3 and rename TLS <= 1.2 variable
TLS v1.3 uses different (incompatible) cipher suite names, so
a new variable is added to set these in roles which require them.

TLS v1.2 and below is still required, but the variable name is
changed to avoid confusion, whilst using the old name as a default
if set by a user deployment.

Change-Id: I9312daa7ca75e484f8aa3857fe43565c89e3f09a