EOL Pike branch
We agreed to EOL Pike as we have no futher plans to provide any patches
to the stable/pike branch.

Due to specifics of releases repo for older branches, we have to
EOL integrated repo and services repos in one shot, so we have to switch
roles to tags that not yet exist.

We switch to noop jobs as the Pike branch relies on external versions
of get-pip.py which no longer exist at the original URL.

Change-Id: I1360988e87d08468d93832226b804dbe666713ca