Add test of used SHAs
We add additional test, that runs bootstarp-ansible.sh and
with that verifies that all roles and collections could be retrieved
from opendev. On top of that it runs extra task that retireves
service_projects and checkout it to the defined SHA.

Test runs only when we change files that we're testing.

Change-Id: I5ad5d4e5b0e59098fae9d185706753397caea6ec