Fix extra facts gathering with tags
When switching to gathering min facts, we introduced common task that
was responsible for gathering specific subset. However we used
include_task and not applied always tag to include. That resulted in
effect that facts were not gathered when playbook run with tags
and leads to failure.

To workaround that we move include to prior facts gathering stanza.

As alternative option for ceph we applied always tag to include.

Change-Id: Ie996262043edb8fb9c7221a5cd98e98a0111ac28
(cherry picked from commit 2492d82da6fe3070528f6155bb918e45d0501809)