Install rally only when rally_install is set
Currently we install rally regardless of user will.
It is default behaviour which is not possible to override.
So we add rally_install variable which defaults to false to control
whether to proceed with rally installation or not.

Change-Id: I325dee6ec071b6675bcfdadaf884ce8c26cf1ae9
(cherry picked from commit 65c4d7ca432ec36915295a6c7167800a8acda5b0)