Update repo verification file URI
Since we're creating a symlink to the U-C file to have a consistent
filename for it, it makes sense to verify this file for simplicity.
With migration to gluster we also should change verification URI
for HaProxy, which we also set to u-c link.

With that we also somehow solve issue when u-c can't be fetched when SHA
is updated, but repo-server role wasn't re-ran.

Closes-Bug: #1943978
Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-repo_server/+/842567
Change-Id: I23ac2ec3246cbae50d30815fbd3bdcc4da7d7089