Run keystone pre-service setup tasks on all keystone hosts first
Some of these tasks, such as setting up ssh keys are required to
be complete on all hosts before the main keystone service deployment
happens.

This change runs a specific tasks file from the os_keystone role
across all of the keystone hosts to do the initial setup, before the
os_keystone role is run again with a serial: setting appropiate for
deploying the actual service code in an HA deployment.

Closes-Bug: 1975718
Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-os_keystone/+/843739
Change-Id: Ib3efe0c43e20904cfbd9b89523e5118413c2282e