Remove usage of haproxy-endpoints role
haproxy-endpoints role was used only to manage galera backends.
To have more consistency and as transition was never finished,
we rollback to using inlcuded task.

Also role usage was a bit complex and hard to read. It also had a
drawback that service can be broken before handler run, which makes
endpoint managing a bit useless in this scenario.

As follow-up we should also drop all "Manage LB" handlers in roles.

Change-Id: I1293776f58a799c9e9dd70f8cd076ed2e2b1bdfe