Do not try to run etcd in serial mode
Etcd static discovery, even in bootstrap mode, tries to wait for all
members to join. Until that systemd unit is in `activating` state which
means that task just timeouts waiting for activation. So etcd can not be
run in serial.

With that we also move etcd installation from zun role to the playbook.
As we still want to run playbook for zun installation/configuration in
serial mode, having etcd in independent play allows us to
do that.

This partially reverts commit Ie996262043edb8fb9c7221a5cd98e98a0111ac28

Change-Id: Ic621a7618b99abb56d24e4dcbb52b495c1b39d57