rabbitmq: default to using TLS for management user interface
The RabbitMQ role defaults change in Yoga to enable the management
UI with TLS/HTTPS. This implicitly disables the HTTP port.

This commit adjusts the HAProxy config to take account of this
change and switch the port used accordingly. The
rabbitmq_management_ssl variable is also set explicitly to ensure
it is defined with appropriate scope.

Change-Id: I5a9f9855aa701d12bc3c9e2e7e9c651ff606c319