Let git choose the branch when cloning a repo
In this path of the code we clone then checkout the required refspec.
The branch name used during the initial clone does not affect the
checkout to the required refspec.

Hardcoding the initial clone to master has worked when the default
branch name for repos has been 'master'. The migration of default
branch names to 'main' breaks this assumption.

By not specifying a branch name in the initial clone we avoid
problems with master/main migration, or with a repo which legitimately
uses a non standard default branch name.

Cherry-Picked from https://review.opendev.org/c/openstack/openstack-ansible-plugins/+/846440

Change-Id: I94a5eb8388c690ee593771815c92e6fada6fb0f4