Remove neutron agents from ironic env.d file
The neutron agents are not required to be installed as part of ironic.
The env.d file required them to be installed in ironic_compute_container
which worked for metal deploys as this is usually the infra node where
the neturon agents are also installed by default. In the case of an
LXC deployment, br-vxlan and br-vlan were required to be wired to
the ironic container to install another copy of the network agents
duplicating those on the infra (or network) hosts.

Also, if seperate hosts are defined in the inventory for network agents,
the ironic installation might fail completely as br-vlan and br-vxlan
(or equivalent) may not be present on the infra nodes at all.

This patch removes the need to install the linuxbridge and ovs network
agents in the same location as the ironic compute service.

Change-Id: Icb87b7bbfea0c2420e37b591496c8b2fa54c096c