Fix dynamic-address-fact gathering with tags
Right now an intermittent issue exists, since setup module does not run
when play is launched with --tags. dynamic-address-fact.yml is included
rather then imported thus tags should be properly applied in include
statement.

Because of that, facts were not gathered, and if facts are expired,
this would lead to changing config and all kind of nasty issues.

So we apply `always` tag to the facts gathering to ensure facts are
in place.

Change-Id: Ib1e3a47acc34ff6d8e6de9555aea59ee8aa244e7