Require usage of service_token_roles
With Yoga we've added `service` role for each service and set it as
`service_token_roles`. For upgrade purposes service_token_roles_required
was set to False, as otherwise services won't be able to comunicate
until upgrade is finished. Now we remove override and require usage
of service_token_roles by default.

Change-Id: I6e57c26dcae1e1470280dc5988903b79f9cb9b16