Run test_minimum_basic_scenario test for cinder
Currently we run only test_server_basic_ops scenario that does launch
VM from ephemeral drive. While it's important thing to test, we still
need to ensure, that volume-based instances does also work properly.
For that we will run test_minimum_basic_scenario when os_cinder
role is running in CI. We also want to minimize amount of test to
reduce time in CI, thus it's not added as a general test.

Change-Id: I037f0eeafae9cd8dc2f65deee1a12f28ed78ad78