Replace usage of which with command
`which` is not a default package for CentOS 9 and Rocky Linux. This
means that we should either take care of installing `which` at very
beginning or use built-in alternative of `command -v`.
Command is part of GNU Bash package and will present on each system
which have bash installed, which is OSA requirements.

`command -v` behaves exactly same way - it provides path as output
and exits with code 1 when binary can't be found. So it's a drop-in
replacement.

Change-Id: I7da7814261d734b96644624870ed7cc2f1173bff