Define a group for ironic_console
It is possible to install nova-serialconsole-proxy alongside
nova-compute to allow IPMI serial-over-lan consoles to be made
available through the openstack cli or through horizon.

The os_nova role needs to have logic to decide which ironic host to
install the nova serial proxy to, so we define an extra ansible group
for this purpose.

Change-Id: I936aa89129aeda56af9d9a1407073fccf00b7170