Update self version fetching mechanism
With release of new setuptools running setup.py --version started
returining extra line. Eventually, running setup.py as a script is not
considered as good practice and we should not rely on it's output much.
Instead, we can leverage importlib.metadata as it should be more
reliable in long run.

Closes-Bug: #1997365
Change-Id: Ife227db4ad25851456e1acc6b4daf198f39d989a