Simplify glance deployment logic
As of today there's no need in glance RPC version negotiation.
Also, glance is being restarted anyway in case of changes in venv
or config inside the role, to addintional restart after the role simply
should not be needed anymore. Restarting service once on changes
should be enough for glance.

Change-Id: Id0edd4205b4e8048d7191b4bcfce83b0442689ea