Define name for all collections in a-r-r
At the moment it's not possible to override collection, since we merge
user and default collections by name. So in case it's needed to
override git repo or user wants to use galaxy, they can only extend
installable collections. At the same time in envs with limited
connectivity it might be impossible to fetch data from third-party repos
and replacement is needed.

So to allow proper overrides we place collection name and source
keys regardless of type.

Change-Id: Ic0b125715527e4f3fba5238c3f7577e83c93e568