Do not run dstat by default
Fixing dstat runtime in gate resulted in a side-effect, that dstat_graph
tries to run even outside of the gates, since GATE_EXIT_RUN_DSTAT
was always defaulted to True. Now we enable dstat only when ZUUL_PROJECT
variable is defined. In it's turn ZUUL_PROJECT is set only in zuul pre-run
tasks for our jobs.

Change-Id: I807985f68e1a090c5b3c055a78c16b79beeec8c5