Prevent bootstrap failure when all roles/collections are overriden
At the moment we assume that scenario when all roles or collections
provided by ansible-role/collection-requirements won't be overwritten.
As if they do, previous set_fact will be skipped and
galaxy_collections_list/clone_roles will remain undefined.

To avoid failure and to allow to fully override roles/collections we
define default for the task that aims to merge default items with
user-provided ones.

Change-Id: I86d44499073f0571a529d81ed3be713996e5f339