Ensure management_address is used instead of ansible_host
ansible_host in deployments is designed to represent SSH address.
In cases, when address that used for SSH is different from management
network, this might lead to undpredictable results on bare metal hosts.

Thus we're ensuring that management (container) address is used
for services to listen on and to interact through.

Related-Bug: #2002645
Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/870113
Change-Id: I8a5e817d024eef5453fa072d8fee5aeca9bed67b