Backport openstack_testing from Yoga
At the moment we're facing issues with tempest failing to install
all required plugins due to not matching constraints with requirements

In order to prevent that we used to fix plugin versions directly in the
tempest role in an ad-hoc manner. However, it's better to use
correct and working versions of plugins for the branch, so SHAs were
backported from Y which should work correctly.

Change-Id: I11f5326724ac1af00166f2d6d86cd3427470fea0