Backport openstack_testing from Yoga
At the moment we're facing issues with tempest failing to install
all required plugins due to not matching constraints with requirements

In order to prevent that we used to fix plugin versions directly in the
tempest role in an ad-hoc manner. However, it's better to use
correct and working versions of plugins for the branch, so SHAs were
backported from Y which should work correctly.

Also according to no objections raised during the meeting [1], we're
disabling upgrade jobs from branches that are currently in EM, since
we don't need to block current development due to EM failures.

[1] https://meetings.opendev.org/irclogs/%23openstack-ansible/%23openstack-ansible.2023-03-07.log.html#t2023-03-07T15:06:59

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-rabbitmq_server/+/875782
Change-Id: I11f5326724ac1af00166f2d6d86cd3427470fea0