Use certbot to generate SSL cert for the external VIP in 'stepca' scenario
This patch uses the certbot functionality to issue a certificate
from a locally installed ACME server on the AIO node, when the
string 'stepca' appears in the scenario.

This allows testing of the certbot code in the haproxy role and the
wider integration with Openstack-Ansible to be tested in CI jobs.

Change-Id: Ide769f54505898630aae67e25b238624ba4f4fdb