Use a map file to select haproxy horizon backend from the base frontend
This is the smallest change to demonstrate the use of a map file
with haproxy. A 'map_reg' type regular expression map is used to
match any traffic arriving at the base frontend and direct it
to the horizon backend, which duplicates the current behaviour.

The map file contents used behave the same as the 'default_backend'
config directive on a frontend.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-haproxy_server/+/876749
Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-haproxy_server/+/878771
Change-Id: I95d0d2523c8452aeb624923b40a55c09d9341ac3