Remove nova-scheduler from ironic compute container_skel
There is no need to have instances of nova-scheduler in the
ironic_compute host group. Nova scheduler will already be running
in the nova_scheduler host group and extra instances of the
scheduler bring confusion when trying to reconfigure or
debug the main nova service.

This patch removes the definition of the scheduler from the
group - for an existing deployment the nova scheduler service
any extra instances should be stopped/disabled/masked with
systemd.

Change-Id: Ie5a94966e0740f41b412ddf9bd79fbc5254829e9