Split compatability matrix for legacy releases
With time amount of operating systems in the table become hardly
managable, thus we split table into 2 parts, current one is going
to contain only currently supported releases, while all EM and EOL
are moved to the legacy one. Likely, with time we might want to split
these even more to have EOL and EM as a separate tables, but it's not
done now due to not having significant simplification of tables.

Change-Id: I4365c00de3561bfb4b1ac0b298ae670c1c00db93