Add documentation for multi-architecture deployments.
This is supported natively by OpenStack-Ansible but the deployment
considerations are subtle and require appropriate overrides by
the deployer and image properties to be applied by end users.

This patch adds a documentation section oriented around the
user story of adding new aarch64 compute nodes to an existing
x86_64 deployment.

Change-Id: Ie517a46b0e5452a78d9a5ad05e9bb778e71398ef