Define individual haproxy services for each console type
The current code in openstack-ansible assumes that only one
console type is active at once, with a special case for when
ironic is deployed to also enable serial consoles.

This does not cover the case when different nova compute nodes may
require different console types, such as spice/novnc for x86_64
and serialconsole for aarch64.

This patch maintains the same external variables (nova_console_tpye
and ironic_console_type) - but makes specific haproxy backends for
each type. This is the first step required to allow multuple console
types to be enabled at the same time for nova compute nodes.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/879069
Change-Id: Ib6f77036639568321d07e9f478c1e087bd9fee91