Fix dstat run in gates
With previous commit [1] a condition was introduced on when to run
dstat. Though, ZUUL_PROJECT env variable used is defined
only in post jobs, while dstat should be started in pre-jobs and then
collected with logs. So we adjust condition to start up dstat in gates
at the first place.

[1] https://opendev.org/openstack/openstack-ansible/commit/0369605a1c560981fbcf4171f0c9e195502fcfe9

Change-Id: I7fd4a846360847496b1db7b505b35b9eff57ebbe