Replace CentOS 9 Jobs with Rockylinux
During PTG there was an agreement to focus more on Rocky testing
comparing to CentOS Stream since these jobs are quite troublesome
lately and not due to our fault. Right now CentOS is completely borked
due to gnupg release backport [1]
With that to unblock CI and to execute PTG decision, we're marking
all CentOS jobs as non-voting and replacing them with Rocky 9 jobs.
We also mark Rocky as voting for upgrade jobs.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2184640

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-galera_server/+/879680
Change-Id: Id431dd8d49f86079b6ca3f0f16718489f60ffa77