Add Yoga upgrade jobs
This patch aims to implement testing of SLURP upgrades. While
2023.1 is the first SLURP release, upgrade from Yoga to 2023.1 is
considered as unofficial SLURP upgrade.

With that we're adding changes to gate-check-commit to be able to
provide source release from which upgrade will be perfromed.

While running upgrade script user is required to provide source
release as input to script for SLURP releases.

Also regular upgrades are now switched to Zed as they supposed to be.

Change-Id: I64e55b2c685782d23bb84e7e9f7c60708c276cc3