Switch ubuntu upgrade jobs to Jammy
Since Zed we're fully supporting jammy as distirbution and thus should
execute upgrades from Zed to 2023.1 using jammy. At the same
time upgrades from Yoga to 2023.1 will remain on focal as jammy
was not supported yet on Yoga.

We're also adding workaround for OVS bug that's present for Jammy and
OVN scenario.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-os_neutron/+/883740
Change-Id: I3c41c756812336d9ed0efec05d602a6781e20b81