Install zuul-provided collections as directories
At the moment we don't really use collection version that is
provided by depends-on. This results in testing different content
that should be. Presumably, this happens because ansible-galaxy
does install from default branch HEAD, while zuul does check-out
to refs when using depends-on. To ensure that git repo state is not
the issue, we switch type to be directory for collections that are
contained in Zuul.

Change-Id: I7a0e2f7702a9f8630ed916523b9f0e6c2bbc202b